Adapted model to existing database

+++ src/main/java/com/utpl/plansvalidator/custom_queries/PlanDocenteQueryResolver.java
@@ -22,7 +22,7 @@ public class PlanDocenteQueryResolver {
     @PersistenceContext
     private EntityManager entityManager;
 
-    public CustomResultList executeQuery(String query, String condition, Long planId) throws JSQLParserException, SQLSyntaxErrorException {
+    public CustomResultList executeQuery(String query, String condition, int planId) throws JSQLParserException, SQLSyntaxErrorException {
         String[] splitQuery = query.toLowerCase().split("from");
         query = String.format(QUERY_WRAPPER, splitQuery[0], splitQuery[1], condition, planId);
         List<String> columnNames = new ArrayList<>();
